+++ src/briefy/leica/events/task.py
@@ -1,0 +1,9 @@
+"""Custom briefy.leica events and subscribers."""
+from briefy.common.event import TaskEvent
+from briefy.leica import logger
+
+
+class LeicaTaskEvent(TaskEvent):
+    """A task on Leica."""
+
+    logger = logger