+++ src/briefy/leica/tools/oneshots/20170419-fix-scout_manager-orders-assignments.py
@@ -1,0 +1,67 @@
+"""Find and set all missing scout managers in Orders and Assignments."""
+from briefy.leica.db import Session
+from briefy.leica.models import Assignment
+from briefy.leica.models import Order
+from briefy.leica.models.job.workflows import SELF_ASSIGN_SCOUT_ID
+from briefy.leica.sync.db import configure
+from briefy.leica.tools import logger
+from sqlalchemy import not_
+
+import transaction
+
+
+def update_orders_scout_manager():
+    """Set scout manager in the orders without."""
+    all_orders = Order.query().filter(
+        not_(
+            Order.state.in_(
+                ['received', 'cancelled']
+            )
+        )
+    ).all()
+
+    no_scout_manager = [o.id for o in all_orders if not o.scout_manager]
+    total = len(no_scout_manager)
+    logger.info('{0} Orders without scout manager will be updated.'.format(total))
+    for number, item in enumerate(no_scout_manager):
+        order = Order.get(item)
+        assignment = order.assignments[-1]
+        assignment_scout = assignment.scout_manager
+        order.scout_manager = assignment_scout if assignment_scout else order.project_manager
+        logger.info('Scout Manager set on Order {0}'.format(order.id))
+        if number % 10 == 0:
+            transaction.commit()
+
+
+def update_assignments_scout_manager():
+    """Set scout manager in the assignments without."""
+    all_assignments = Assignment.query().filter(
+        not_(
+            Assignment.state.in_(
+                ['pending', 'cancelled']
+            )
+        )
+    ).all()
+
+    no_scout_manager = [o.id for o in all_assignments if not o.scout_manager]
+    total = len(no_scout_manager)
+    logger.info('{0} Assignments without scout manager will be updated.'.format(total))
+    for number, item in enumerate(no_scout_manager):
+        assignment = Assignment.get(item)
+        order = assignment.order
+        order_scout = order.scout_manager
+        if assignment.pool_id:
+            assignment.scout_manager = SELF_ASSIGN_SCOUT_ID
+        else:
+            assignment.scout_manager = order_scout if order_scout else order.project_manager
+        logger.info('Scout Manager set on Assignment {0}'.format(assignment.id))
+        if number % 10 == 0:
+            transaction.commit()
+
+
+if __name__ == '__main__':
+    session = configure(Session)
+    update_assignments_scout_manager()
+    transaction.commit()
+    update_orders_scout_manager()
+    transaction.commit()